Task creation now accounts for completed tasks
If the taskList is created and there are no completed tasks, the
taskList will not contain the completed line. If the tasklist is created
and it contains completed tasks, then they'll be after the line and
marked.

Finally, this added the addition and removal of the completed line when
the tasks are moved around.