List now saves to memory (bug)
There's a bug where when first retrieving the json it's possible
for tasks to be in the right place, but not checked properly.
Tried to add a subscription to the event to update the checkboxes
after the fact, but that didn't work. Tried to make it all wait for the
json to be loaded, but that didn't help, tried to see if the DOM
was loading too slowly, but that wasn't it either.

Will come back later in the hopes that this won't affect the final task
list.