Project page creation started. Many bugs...
There's an issue where jsonProjectLoad is not always seeing that there
are projects in memory and will then create new projects. This is an
issue because then the ID that gets passed along isn't correct.

There's also the issue of the completedLine being a hard requirement
when creating the taskList.