finishedTaskListUpdate & projectpage creation bugs
Both of these bugs showed up as I moved to have an empty project to
check empty state.

finishedTaskListUpdate is a bug when trying to click on the page for
creating a project page that is supposed to be empty.

projectpage creation bug is when a page should have a project, but is
empty.