Single event listener on taskList
Change and click events were conflicting, so the taskList event listener
now has only one event listener and a function to differenciate what
was clicked. Function stayed the same for the input, but now the
button (a or chevron) has a proper PubSub announcing the ID of the
clicked task. Next step is to use it to call task-view.