This being a learning project, we stop here ATM.
I started adding buttons to task-view to edit and delete the tasks,
but I feel like I'd learn more by continuing the Odin Project.

As such, the groundwork is laid for the buttons, but the interactions
need to be created. If you read this at a later time, good you came back!